refactor: unify argmax index on u32 across CPU + CUDA
Settles the argmax index type by aligning the whole library on u32, instead
of the speculative u32→i32 widening (1db15da) that this commit reverts.

Rationale:
- The only real downstream consumer, tensor4all/omeinsum-rs, is u32-native
  throughout its own `Algebra::Index` and reads `get_argmax()` into a
  `Vec<u32>`. The i32 widening was source-breaking for it (6× E0308); u32
  keeps this PR non-breaking and within 0.2.x.
- Argmax values are array k-indices — naturally unsigned. GPUs support u32
  natively (the i32 was a convention, not a hardware constraint).
- The sole motivation for signed was a future "-1 = no edge" sentinel, which
  is deferred; the shipped semantics already canonicalize a no-contribution
  cell to the seed 0 (u32-compatible), and a future no-edge marker would use
  u32::MAX.

Changes:
- CPU core + Python binding: revert 1db15da → argmax index back to u32
  (GemmWithArgmax / MatWithArgmax / get_argmax / tropical_backward_* / the
  TropicalWithArgmax<Index = u32> bound). scalar.rs untouched (its u32/i32 are
  scalar element types, not indices).
- CUDA: `ArgmaxIndex = i32` → `u32`; the alias propagates to argmax_to_host /
  into_parts / from_cuda_slices / the backward signatures. gpu_mat
  to_mat_with_argmax now assigns the u32 buffer directly (no boundary cast).
  Backward GPU kernels take `&CudaSlice<ArgmaxIndex>`.
- Kernels (tropical_gemm.cu): all argmax buffers are now `unsigned int*` —
  7 GEMM-with-argmax outputs and the 2 backward inputs (read via an explicit
  `(int)` cast so the `k >= 0 && k < K` guard is byte-for-byte unchanged).

Result: one argmax index type (u32) across CPU, CUDA, Python, and the
downstream consumer; no boundary conversions. CUDA's internal i32 is gone, so
the CPU/CUDA asymmetry is resolved.

Verified locally: cargo test -p tropical-gemm (274 + 24, 0 fail); clippy
--all-targets clean on tropical-gemm and tropical-gemm-cuda;
PYO3_USE_ABI3_FORWARD_COMPATIBILITY=1 cargo check -p tropical-gemm-python
clean; cargo check -p tropical-gemm-cuda --all-targets clean; and omeinsum-rs
--features tropical-kernels compiles unmodified against this branch. Full
NVRTC build + GPU test suite to be validated on HKUST-GZ HPC (A40).

+++ crates/tropical-gemm-cuda/kernels/tropical_gemm.cu
@@ -549,7 +549,7 @@ extern "C" __global__ void KERNEL_NAME(                                        \
     const float* __restrict__ A,                                               \
     const float* __restrict__ B,                                               \
     float* __restrict__ C,                                                     \
-    int* __restrict__ argmax,                                                  \
+    unsigned int* __restrict__ argmax,                                         \
     int M, int N, int K                                                        \
 ) {                                                                            \
     const int BLOCK_SIZE_M = 64;                                               \
@@ -668,7 +668,7 @@ extern "C" __global__ void KERNEL_NAME(                                        \
     const double* __restrict__ A,                                              \
     const double* __restrict__ B,                                              \
     double* __restrict__ C,                                                    \
-    int* __restrict__ argmax,                                                  \
+    unsigned int* __restrict__ argmax,                                         \
     int M, int N, int K                                                        \
 ) {                                                                            \
     const int BLOCK_SIZE_M = 32;                                               \
@@ -787,7 +787,7 @@ extern "C" __global__ void KERNEL_NAME(                                        \
     const int* __restrict__ A,                                                 \
     const int* __restrict__ B,                                                 \
     int* __restrict__ C,                                                       \
-    int* __restrict__ argmax,                                                  \
+    unsigned int* __restrict__ argmax,                                         \
     int M, int N, int K                                                        \
 ) {                                                                            \
     const int BLOCK_SIZE_M = 64;                                               \
@@ -906,7 +906,7 @@ extern "C" __global__ void KERNEL_NAME(                                        \
     const long long* __restrict__ A,                                           \
     const long long* __restrict__ B,                                           \
     long long* __restrict__ C,                                                 \
-    int* __restrict__ argmax,                                                  \
+    unsigned int* __restrict__ argmax,                                         \
     int M, int N, int K                                                        \
 ) {                                                                            \
     const int BLOCK_SIZE_M = 32;                                               \
@@ -1023,7 +1023,7 @@ extern "C" __global__ void KERNEL_NAME(                                        \
 #define TROPICAL_BACKWARD_A(KERNEL_NAME, TYPE, ATOMIC_ADD)                     \
 extern "C" __global__ void KERNEL_NAME(                                        \
     const TYPE* __restrict__ grad_c,                                           \
-    const int* __restrict__ argmax,                                            \
+    const unsigned int* __restrict__ argmax,                                   \
     TYPE* __restrict__ grad_a,                                                 \
     int M, int N, int K                                                        \
 ) {                                                                            \
@@ -1031,7 +1031,7 @@ extern "C" __global__ void KERNEL_NAME(                                        \
     int total = M * N;                                                         \
     if (idx < total) {                                                         \
         int i = idx % M;                                                       \
-        int k = argmax[idx];                                                   \
+        int k = (int)argmax[idx];                                              \
         if (k >= 0 && k < K) {                                                 \
             ATOMIC_ADD(&grad_a[i + k * M], grad_c[idx]);                       \
         }                                                                      \
@@ -1041,7 +1041,7 @@ extern "C" __global__ void KERNEL_NAME(                                        \
 #define TROPICAL_BACKWARD_B(KERNEL_NAME, TYPE, ATOMIC_ADD)                     \
 extern "C" __global__ void KERNEL_NAME(                                        \
     const TYPE* __restrict__ grad_c,                                           \
-    const int* __restrict__ argmax,                                            \
+    const unsigned int* __restrict__ argmax,                                   \
     TYPE* __restrict__ grad_b,                                                 \
     int M, int N, int K                                                        \
 ) {                                                                            \
@@ -1049,7 +1049,7 @@ extern "C" __global__ void KERNEL_NAME(                                        \
     int total = M * N;                                                         \
     if (idx < total) {                                                         \
         int j = idx / M;                                                       \
-        int k = argmax[idx];                                                   \
+        int k = (int)argmax[idx];                                              \
         if (k >= 0 && k < K) {                                                 \
             ATOMIC_ADD(&grad_b[k + j * K], grad_c[idx]);                       \
         }                                                                      \
@@ -1118,7 +1118,7 @@ extern "C" __global__ void KERNEL_NAME(                                        \
     const float* __restrict__ A,                                               \
     const float* __restrict__ B,                                               \
     float* __restrict__ C,                                                     \
-    int* __restrict__ argmax,                                                  \
+    unsigned int* __restrict__ argmax,                                         \
     int M, int N, int K,                                                       \
     int strideA, int strideB, int strideC                                      \
 ) {                                                                            \
